Updated snowflake extension to explicitly type integer/float/boolean for json columns when generating SQL.   (#2928)
* Updated snowflake extension to explicitly type integer/float/boolean for json columns when generating SQL.

In the past when you generate sql queries with semistructured data in snowflake, there was no type given to columns other than dates and strings.   This causes snowflake's type system to define the type of these results to be "variant" and therefore is not consistent with the pure types.   This update causes explicit types for integer, float, and boolean.   It does not define types for other primitives (e.g. decimal) because there's no way to know what the x and y will be in numeric(x,y).

* Replace empty line that was removed.

Replace empty line that was removed .